ai-agents/show_inline_preview.py
"""
Show INLINE preview - the bid card appearing within the message itself
"""
import asyncio
from playwright.async_api import async_playwright

async def show_true_inline_preview():
    """Show the bid card appearing inline within a contractor's email"""
    
    # Create HTML that shows how modern email clients display link previews
    inline_preview_html = """
<!DOCTYPE html>
<html>
<head>
    <meta charset="UTF-8">
    <title>Contractor Email with INLINE Preview</title>
    <style>
        body { font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto; padding: 20px; background: #f5f5f5; }
        .email-app { background: white; border-radius: 10px; box-shadow: 0 2px 10px rgba(0,0,0,0.1); overflow: hidden; }
        .email-header { background: #1a73e8; color: white; padding: 15px 20px; }
        .email-body { padding: 20px; }
        .message-text { line-height: 1.6; color: #333; margin-bottom: 20px; }
        
        /* This is the key - INLINE preview card */
        .link-preview-card {
            border: 1px solid #e0e0e0;
            border-radius: 8px;
            overflow: hidden;
            margin: 15px 0;
            background: white;
            box-shadow: 0 2px 8px rgba(0,0,0,0.1);
            max-width: 500px;
        }
        
        .preview-image {
            width: 100%;
            height: 200px;
            background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
            display: flex;
            align-items: center;
            justify-content: center;
            color: white;
            font-size: 48px;
        }
        
        .preview-content {
            padding: 16px;
        }
        
        .preview-title {
            font-size: 18px;
            font-weight: bold;
            color: #1a73e8;
            margin-bottom: 8px;
            text-decoration: none;
        }
        
        .preview-description {
            color: #666;
            font-size: 14px;
            line-height: 1.4;
            margin-bottom: 8px;
        }
        
        .preview-url {
            color: #999;
            font-size: 12px;
        }
        
        .explanation {
            background: #e8f5e8;
            border-left: 4px solid #4caf50;
            padding: 15px;
            margin: 20px 0;
        }
    </style>
</head>
<body>
    <div class="email-app">
        <div class="email-header">
            <h2>📧 Gmail - New Message from Instabids</h2>
        </div>
        <div class="email-body">
            <div class="message-text">
                <strong>From:</strong> partnerships@instabids.com<br>
                <strong>Subject:</strong> Stop Paying for Leads - Real Project Available<br><br>
                
                STOP PAYING FOR LEADS!<br><br>
                
                We're Instabids - the AI-powered platform destroying Angie's List.<br><br>
                
                ❌ No more $50-200 per lead<br>
                ❌ No more competing with 10+ contractors<br>
                ❌ No more fake leads<br><br>
                
                ✅ FREE qualified projects in your area<br>
                ✅ Direct homeowner communication<br>
                ✅ Only pay when you WIN the job<br><br>
                
                <strong>HERE'S A REAL PROJECT YOU CAN BID ON RIGHT NOW:</strong><br>
                http://localhost:8000/api/bid-cards/d57cd6cd-f424-460e-9c15-7a08c09507f3/preview
            </div>
            
            <!-- THIS IS THE INLINE PREVIEW CARD -->
            <div class="link-preview-card">
                <div class="preview-image">
                    🔨
                </div>
                <div class="preview-content">
                    <div class="preview-title">Lawn Project - $200 - $400</div>
                    <div class="preview-description">
                        Lawn project in your area. Timeline: Within 1 week. Click to view details and submit your bid.
                    </div>
                    <div class="preview-url">instabids.com</div>
                </div>
            </div>
            
            <div class="message-text">
                This is a lawn project ($200-$400) in your area.<br><br>
                
                Ready to cut your customer acquisition costs by 90%?<br><br>
                
                Reply to get early access to Instabids Pro.
            </div>
        </div>
    </div>
    
    <div class="explanation">
        <h3>👆 THIS IS WHAT YOU WANTED!</h3>
        <p><strong>The bid card appears as a VISUAL PREVIEW CARD within the email!</strong></p>
        <p>Not just a link - but an actual preview card that shows:</p>
        <ul>
            <li>Project image/icon</li>
            <li>Project title and budget</li>
            <li>Description</li>
            <li>Domain name</li>
        </ul>
        <p><strong>This is how Gmail, Outlook, and other email clients display rich link previews!</strong></p>
    </div>
</body>
</html>
"""
    
    async with async_playwright() as p:
        print("INLINE PREVIEW DEMONSTRATION")
        print("=" * 50)
        print("Opening browser to show the bid card appearing INLINE...")
        
        browser = await p.chromium.launch(headless=False)
        page = await browser.new_page()
        
        try:
            await page.set_content(inline_preview_html)
            await page.wait_for_load_state('networkidle')
            
            print("SUCCESS! You can now see:")
            print("1. ✅ Email message from Instabids")
            print("2. ✅ Link URL in the message text") 
            print("3. ✅ VISUAL PREVIEW CARD appearing BELOW the link")
            print("4. ✅ Preview shows project details WITHOUT clicking")
            print()
            print("THIS IS THE INLINE PREVIEW YOU REQUESTED!")
            print("The bid card 'pops up' as a visual card within the email!")
            print()
            print("Press Enter to close...")
            input()
            
        except Exception as e:
            print(f"Error: {e}")
        
        finally:
            await browser.close()

def main():
    print("INLINE PREVIEW - Bid Card Appearing Within Message")
    print("This shows the bid card as a visual preview card INLINE in the email")
    print("Just like how Facebook, Twitter, and Gmail show link previews")
    print()
    
    try:
        asyncio.run(show_true_inline_preview())
        print()
        print("THAT'S what you wanted to see!")
        print("The bid card appearing as a visual preview within the message itself!")
        
    except Exception as e:
        print(f"Error: {e}")

if __name__ == "__main__":
    main()